% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{batch_correction}
\alias{batch_correction}
\title{Batch Correction}
\usage{
batch_correction(spe, email_address, BPPARAM = NULL)
}
\arguments{
\item{spe}{spe object}

\item{email_address}{email address to send updates}

\item{BPPARAM}{parallelization param}
}
\description{
Batch Correction
}
\examples{
batch_correction(spe, t0jone12@louisville.edu)

}
