% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_manage.R
\name{fcs_manage}
\alias{fcs_manage}
\title{Manage FCS files into an all dataset}
\usage{
fcs_manage(
  path_to_files = "data/raw_data/",
  path_to_csv = "data/derived_data/"
)
}
\arguments{
\item{path_to_files}{path to location of fcs files}

\item{path_to_csv}{path to location to write csv file}
}
\value{
file written to path_to_csv
}
\description{
Manage FCS files into an all dataset
}
\examples{
fcs_mange()

}
