% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcs_to_df.R
\name{read_clean_fcs}
\alias{read_clean_fcs}
\title{Read and CLean FCS Files}
\usage{
read_clean_fcs(file, path_to_files)
}
\arguments{
\item{file}{}
}
\value{
cleaned dataframe with fcs info
}
\description{
Read and CLean FCS Files
}
\examples{

read_clean_fcs("path/to/file.fcs")
}
